third_party/WebKit/LayoutTests/http/tests/push_messaging/permission-state-denied-in-document.html

<!DOCTYPE html>
<html>
<head>
<title>Push API: permissionState resolves with denied when permission is denied</title>
<link rel="manifest" href="resources/push_manifest.json">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../serviceworker/resources/test-helpers.js"></script>
</head>
<body>
<script>
async_test(function(test) {
    var workerUrl = 'resources/empty_worker.js';
    var workerScope = 'resources/scope/' + location.pathname;
    service_worker_unregister_and_register(test, workerUrl, workerScope)
        .then(function(serviceWorkerRegistration) {
            // If running manually, deny permission when prompted.
            if (window.testRunner)
                testRunner.setPermission('push-messaging', 'denied', location.origin, location.origin);
            return serviceWorkerRegistration.pushManager.permissionState({userVisibleOnly: true});
        })
        .then(function(permissionStatus) {
            assert_equals(permissionStatus, 'denied');
            return service_worker_unregister_and_done(test, workerScope);
        })
        .catch(unreached_rejection(test));
}, 'Push API: permissionState resolves with denied when permission is denied');
</script>
</body>
</html>
